Minor cleanups and tweaks
Fix a couple of minor typos to get PRRTE into the path and
ld_library_path. Modify the PRRTE install to take a branch of the rhc54
repo so we can test the refactored version. Correct the MCA param to
point PRRTE to the hostfile

What we really need is:

* a way to pass an option that specifies the repo/branch or tarball to
be used for building PMIx and PRRTE

* an option that directs docker to reresh and rebuild the PMIx and/or
PRRTE installations. Only way I could trigger it was to modify the
Dockerfile.ssh - would be nice if there was some other way to do it

* if we come in as mpiuser, we cannot go into prrte/examples and compile
them as we don't have write permissions to that area. However, if iwe
compile an example and store the executable somewhere else, then the
other nodes cannot find it. Is there some way to either make the install
directory writeable by "mpiuser" or have a directory that is writeable
and is NFS mounted?

Signed-off-by: Ralph Castain <rhc@pmix.org>